requirements.txt
streamlit
neo4j
pandas
sentence-transformers
faiss-cpu
openai
python-dotenv